Improved structure of scripts
