+++ scripts/meta/version.js
