+++ scripts/transform/index.js
@@ -1,0 +1,13 @@
+import { transformJs } from "./transformJs.js";
+import { transformPackageJson } from "./transformPackageJson.js";
+import { transformReadmeMd } from "./transformReadmeMd.js";
+import { transformWebTypes } from "./transformWebTypes.js";
+import { processTagNamesNaive } from "./helpers.js";
+
+export {
+  transformJs,
+  transformPackageJson,
+  transformReadmeMd,
+  transformWebTypes,
+  processTagNamesNaive,
+};
